chore(ci/test/dev): refresh project deps + best practices
Context: I wanted to contribute and locally things were broken for me at
least. Not able to run the tests + a few best practices I wanted to
contribute to.

This commit includes:
- fix the yarn version for all environements easily, if you want to upgrade it you just have to
upgrade bin/yarn-$version and remove the previous one. Basically this
commit will now ensure that anyone or any tool using the project will
always use the same yarn version
- fix local tests, I had a hard time running the tests locally, I had to
fix a few of them. Unsure why it was working in travis though. Some
tests were also not precise enough, for example the export test was
checking an email value but digging further that email value was
duplicated in the dataset..
- I updated the nodejs version
- I updated most dependencies and removed ones that were not used at all
- I will now plug renovate on the repo for dependencies to update easily