feat(deleteinidicespattern): pattern based indice deletion
This command adds a way to delete multiple indices using a pattern. A
recent customer had hundreds of indices and thus wanted to delete a lot
of them, but not all of them at once. This is super tricky because you
cannot delete a replica, you first have to remove the replica from the
main index and then you can delete the replica.

This command deleteindicespattern aims at making this easier.